# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dokkai API
  version: 0.1.0
  description: |
    **Dokkai** turns a codebase into a graph-aware vector database that local AI
    agents can read, search and document.

    This document describes the HTTP API. The running server also serves an
    interactive Swagger UI at `/docs` and ReDoc at `/redoc`, both generated from
    `/openapi.json`; this file is an enriched, portable copy (it additionally
    documents the SSE chat stream and the async-job payloads, which the
    auto-generated spec cannot express well).

    ### Typical flow
    0. `POST /auth/login` — every route below is authenticated (see the
       `Auth` tag); log in (default `admin`/`admin` on first boot) and send
       the returned token as `Authorization: Bearer <token>` on every
       subsequent request.
    1. `POST /config/llm` — configure the chat model (or set `OLLAMA_CHAT_MODEL`
       so it is auto-configured on startup). Requires the `admin` role.
    2. `POST /instances/pipeline` — ingest a repository. Returns a `job_id`.
    3. `GET /instances/jobs/{job_id}` — poll until `status = succeeded`.
    4. `POST /chat` — ask questions; the answer streams back over SSE.

  license:
    name: License TBD
servers:
  - url: http://localhost:8000
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /chat:
    post:
      tags: [Chat]
      summary: Chat with the codebase (SSE stream)
      description: |
        Retrieves the connected code context for `message` and streams an
        audience-specific answer over **Server-Sent Events**
        (`Content-Type: text/event-stream`).

        The stream emits these event types, in order:

        | event    | `data` payload |
        |----------|----------------|
        | `sources`| JSON array of retrieved chunks (see `SourceChunk`). Sent once, first. |
        | `token`  | A JSON-encoded string — one streamed chunk of the answer. Repeated. |
        | `done`   | `{ "conversation_id": "...", "answer": "<full answer>" }`. Sent once, last. |
        | `error`  | A JSON-encoded error string. Terminates the stream. |

        `project_name` **must** match the `project_name` of an ingested repo
        (the repo's top-level folder name). A non-matching value returns an
        empty `sources` array and an ungrounded answer.

        Note: curl needs `-N` to disable buffering and see tokens as they arrive.

        Requires the `admin` or `user` role.
      operationId: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            examples:
              developer:
                summary: Developer question
                value:
                  message: how does the alarm flow work?
                  project_name: your-repo
                  audience: developer
      responses:
        "200":
          description: |
            An SSE stream. Each event is `event: <type>\ndata: <json>\n\n`.
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                stream:
                  summary: Example event sequence
                  value: |
                    event: sources
                    data: [{"entity_type":"Class","name":"AlarmValidationUseCase","qualified_name":"src.alarms.AlarmValidationUseCase","file_path":"src/alarms/validation.ts","start_line":10,"end_line":80,"chunk_text":"...","description":"Validates a detection and notifies clients.","score":0.77,"hop":0,"via":"seed (matched query)"}]

                    event: token
                    data: "The alarm flow"

                    event: token
                    data: " works by ..."

                    event: done
                    data: {"conversation_id":"b0dc2159-5b52-4b73-a7ca-9f094496d64b","answer":"The alarm flow works by ..."}
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chat/conversations:
    get:
      tags: [Chat]
      summary: List conversations
      description: Requires any authenticated role.
      operationId: listConversations
      responses:
        "200":
          description: All stored conversations (persisted in Postgres).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chat/conversations/{conversation_id}:
    get:
      tags: [Chat]
      summary: Get a conversation
      description: Requires any authenticated role.
      operationId: getConversation
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full conversation history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags: [Chat]
      summary: Delete a conversation
      description: Requires the `admin` or `user` role.
      operationId: deleteConversation
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Conversation deleted
                  conversation_id:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ---------------------------------------------------------------------------
  # Config
  # ---------------------------------------------------------------------------
...
components:
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Request body failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
    Unauthorized:
      description: Missing, invalid or expired bearer token.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missing:
              summary: No Authorization header
              value:
                detail: "missing bearer token"
            invalid:
              summary: Unknown or expired token
              value:
                detail: "invalid or expired session"
    Forbidden:
      description: Authenticated, but the user's role cannot perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            detail: "role 'viewer' cannot perform this action"
    ServiceUnavailable:
      description: >-
        Postgres is unreachable. Auth itself is Postgres-backed, so this can
        surface on *any* non-public route (the bearer token can't be
        resolved), not just the ones that store data in Postgres.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Opaque session token returned by `POST /auth/login`, sent as
        `Authorization: Bearer <token>`. 30-day TTL, revoked by
        `POST /auth/logout` or a password change. See the Auth tag.
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      example:
        detail: job not found
...
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf: [{ type: string }, { type: integer }]
              msg: { type: string }
              type: { type: string }
...
    ChatRequest:
      type: object
      required: [message, project_name]
      properties:
        message:
          type: string
          example: "how does the alarm flow work?"
        project_name:
          type: string
          description: Must match an ingested repo's project name.
          example: your-repo
        audience:
          type: string
          enum: [developer, manager, customer]
          default: developer
        conversation_id:
          type: string
          nullable: true
          description: Continue an existing conversation, or omit to start a new one.
        top_k:
          type: integer
          minimum: 1
          maximum: 50
          default: 8
          description: Number of seed chunks to retrieve.
    SourceChunk:
      type: object
      description: A retrieved code chunk, as emitted in the `sources` SSE event.
      properties:
        entity_type: { type: string, example: Class }
        name: { type: string }
        qualified_name: { type: string }
        file_path: { type: string }
        absolute_path: { type: string, description: "Absolute path on disk, as persisted at ingest time (empty if the project was ingested before this field existed)." }
        start_line: { type: integer, nullable: true }
        end_line: { type: integer, nullable: true }
        chunk_text: { type: string }
        description: { type: string, description: Tier-2 micro-description (may be empty). }
        score: { type: number, nullable: true }
        hop: { type: integer, description: "0 = seed, 1 = direct neighbor, ..." }
        via: { type: string, description: "Provenance, e.g. 'calls ← registerRoutes'." }
    ConversationSummary:
      type: object
      properties:
        conversation_id: { type: string }
        project_name: { type: string }
        audience: { type: string }
        title:
          type: string
          nullable: true
          description: Derived from the first user message at creation; null for conversations created before this field existed (UI falls back to the preview).
        message_count: { type: integer }
        last_message_preview: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
    ConversationMessage:
      type: object
      properties:
        role: { type: string, enum: [user, assistant, system] }
        content: { type: string }
        timestamp: { type: string }
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceChunk"
    Conversation:
      type: object
      properties:
        conversation_id: { type: string }
        project_name: { type: string }
        audience: { type: string }
        title:
          type: string
          nullable: true
          description: Derived from the first user message at creation; null for conversations created before this field existed (UI falls back to the preview).
        created_at: { type: string }
        updated_at: { type: string }
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ConversationMessage"
...